[#96] Update to 1.17 (#106)
* feat: 1.17 compatibility

* feat: add nms support for v1_17_R1

* chore: add java 16 to actions

* chore: add 1.17 to buildtools.sh

* chore: add github action debug

* chore: fix build-script

* chore: update release job to java 16

* fix: github actions cache

* fix: cache mojang mappings in .m2

* feat: finalize HeightAccessor

* feat: add 1.17 default config

* chore: add deepslate to 1.17 default config

* fix: proximity HideCondition mismatch on y < 0 with default config

* chore: remove imprex BlockPos from NmsManagers

+++ orebfuscator-common/src/main/java/net/imprex/orebfuscator/config/ProximityConfig.java
@@ -26,7 +26,7 @@ public interface ProximityConfig {
 	boolean useFastGazeCheck();
 	void useFastGazeCheck(boolean fastGaze);
 
-	Map<Material, Short> hiddenBlocks();
+	Map<Material, Integer> hiddenBlocks();
 
 	Map<Material, Integer> randomBlocks();
